Replace int usage with uint32_t in multiple files
Rebuilt the loop control statements in several classes such as SCBE_Call, Parser_Expression, and Parser_Var among others to use the uint32_t data type instead of int. This change promotes better compatibility and general better practices as it offers a better range than int. Also, converted some static_cast<int> calls to uint32_t or int32_t as needed.

+++ src/ModuleDepManager.cpp
@@ -693,7 +693,7 @@ bool ModuleDepManager::execute()
         for (auto d : val->moduleDependencies)
             depNames.insert(d->module->name);
 
-        for (int i = 0; i < static_cast<int>(dep.size()); i++)
+        for (uint32_t i = 0; i < dep.size(); i++)
         {
             auto d = dep[i];
             if (!d->module)
